refactor(digest): do not rely on toString for digest
Digest was coupled to RopesNodes#toString.
This feature is only for debugging purpose and do not offer any
guarentee about its output.

RopesNodes#toList outputs a linear representation of a the tree
such that two distinct trees with the same items have the same
linear representation.
LogootSRopes#digest is computed over this linear representation.