feat(idfactory): Edit createBetweenPosition() to compare identifiers tuple by tuple
Previously, createBetweenPosition() was comparing the two identifiers one value by one.
As soon as a gap wide enough was detected between two values, the comparison was stopped
and a new identifier was created using the common prefix of the two identifiers and a new tuple.
As a result, it generated identifiers of different size respecting the following pattern :

  [number, ]* randomNumber, siteId, clock, position

This could lead to some scenarios in which we tried to compare a random number to a site id,
two values which are not necessarily of the same type.
To prevent this, we now compare the identifiers tuple by tuple,
which results in the generation of identifiers according to the pattern :

  [randomNumber, siteId, clock, position]+