fix(renamablereplicablelist): fix clock value after  renameLocal()
The clock value was actually increased before generating the new identifiers but not afterward.
It leads to cases in which the node would later generate another identifier, but with the same <replicaNumber, clock, offset>.
This bug was breaking the uniqueness property of the identifiers, leading to bugs.
Now increase correctly the clock value after performing the renaming.