fix(renamingmap): fix reverseRenameId()
As illustrated by the test introduced in a previous commit, the transformation applied by reverseRenameId() to ids such as
        id = closestPredecessorOfNewFirstId + tail with firstId < tail
may conflict with the transformation applied to ids such as
        id' = closestPredecessorOfNewFirstId + closestPredecessorOfFirstId + tail' with tail < tail'
The conflict results in the inversion of the order relation between the two given ids.
To solve it, add a MAX_TUPLE to the resulting id.